libsemanage: always write kernel policy when check_ext_changes is specified
For the use case of rebuilding the policy afte package updates, we need
this conditional rebuild to always do at least the do_write step,
because the various semanage dbs may have also changed content relative
to the current binary policy. As this step is itself relatively fast, we
can do it unconditionally.

Fixes: 286a679fadc4 ("libsemanage: optionally rebuild policy when modules are changed externally")
Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>